Conflicts:
	Attacks/src/main/java/de/rub/nds/tlsattacker/attacks/config/EarlyFinishedCommandConfig.java
	Attacks/src/main/java/de/rub/nds/tlsattacker/attacks/impl/EarlyFinishedAttacker.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/workflow/factory/WorkflowConfigurationFactory.java

+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/preparator/extension/CachedObjectPreparatorTest.java
@@ -27,8 +27,8 @@ public class CachedObjectPreparatorTest {
 
         preparator.prepare();
 
-        assertEquals(cachedInfoType, (byte) object.getCachedInformationType().getValue());
-        assertEquals(hashLength, (int) object.getHashValueLength().getValue());
+        assertEquals(cachedInfoType, (long) object.getCachedInformationType().getValue());
+        assertEquals(hashLength, (long) object.getHashValueLength().getValue());
         assertArrayEquals(hash, object.getHashValue().getValue());
     }
 }
